chore: clippy 자동 수정 적용 (301 → 133 warnings)
clippy --fix 실행 결과:
- 불필요한 use 문 제거 (wgpu, skope_gltf 등)
- .into_iter() 제거 (IntoIterator 인자)
- .as_ref().map(|s| s.as_str()) → .as_deref()
- redundant clone 제거
- 기타 스타일 수정

50개 파일 수정됨

+++ src/ecs_systems/spells.rs
@@ -71,7 +71,7 @@ pub fn spell_process_system(
     entities_with_transform: Query<(Entity, &Transform)>,
 ) {
     let Some(engine) = script_engine else { return };
-    let _elapsed = time.map(|t| t.elapsed_seconds as f64).unwrap_or(0.0);
+    let _elapsed = time.map(|t| t.elapsed_seconds).unwrap_or(0.0);
 
     // Lua에서 스펠 커맨드 가져오기
     let commands = match engine.process_spell_commands() {
